refactor: move DifficultyPage content to common/WordListPage
